requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
billiard==4.2.1
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
distlib==0.3.9
Django==4.2.16
django-celery-beat==2.7.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
exceptiongroup==1.2.2
factory-boy==3.3.0
Faker==25.9.2
fakeredis==2.23.3
filelock==3.16.1
firebase-admin==6.6.0
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==22.0.0
httplib2==0.22.0
identify==2.6.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mirakuru==2.5.3
msgpack==1.1.0
multidict==6.1.0
nodeenv==1.9.1
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
port-for==0.7.4
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.8.0
pytest-env==1.1.5
pytest-factoryboy==2.7.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-order==1.3.0
pytest-redis==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==5.0.7
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
sentry-sdk==2.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tomli==2.2.1
twilio==9.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.26.5
wcwidth==0.2.13
whitenoise==6.7.0
yarl==1.18.3
